moves Scroll UP button to Layout
