+++ src/UI/UpButton.css
@@ -2,8 +2,9 @@
   display: none
 }
 
-@media (min-width: 768px) {
+@media (min-width: 844px) {
   .up-button {
+    display: block;
     font-family:'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
     padding: 0.7rem 1rem;
     margin: 0.5rem;
