+++ src/pages/Round.module.css
@@ -32,13 +32,6 @@
   }
 }
 
-@media (min-width: 1000px) {
-  .up-button--round {
-    display: block;
-    right: 6%;
-  }
-}
-
 .error {
   text-align: center;
   margin-bottom: 4rem;
